Added confirmation dialogs for deleting stages and SWIs
SWI-107 #done

+++ swi-app/src/app/modules/swi-manager/components/swi-manager-screen/swi-manager-screen.component.ts
@@ -1,8 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewContainerRef } from '@angular/core';
 import { ActivatedRoute, Router } from "@angular/router";
 import { SWIHeader } from "../../../../models/app.models";
 import { SWIFileService } from "../../../../services/swi-file.service";
 import { ToastsManager } from "ng2-toastr";
+import { Overlay } from "angular2-modal";
+import { Modal } from "angular2-modal/plugins/bootstrap";
 
 @Component({
   selector: 'swi-swi-manager-screen',
@@ -13,13 +15,18 @@ export class SwiManagerScreenComponent implements OnInit {
 
   swi: SWIHeader;
   title: string;
+
   constructor(
+    public overlay: Overlay,
+    public vcr: ViewContainerRef,
+    public modal: Modal,
     private route: ActivatedRoute,
     private router: Router,
     private swiFileService: SWIFileService,
     private toast: ToastsManager
   ) {
     this.swi = this.route.snapshot.data['swi'];
+    overlay.defaultViewContainer = vcr;
     this.title = "SWI Manager - " + this.swi.title;
   }
 
@@ -39,12 +46,30 @@ export class SwiManagerScreenComponent implements OnInit {
   }
 
   deleteSWI() {
-    this.swiFileService.deleteSWI(this.swi.id).then(((delSwi: SWIHeader) => {
-      this.toast.warning(this.swi.title + ' was deleted!', "Successfully Deleted");
-      this.navBack();
-    })).catch((err) => {
-      this.toast.error("Could not delete the SWI", "Delete failed");
-    });
+
+    this.modal.confirm()
+      .size('lg')
+      .isBlocking(true)
+      .showClose(false)
+      .keyboard(27)
+      .titleHtml('<h5>Confirm Delete SWI</h5>')
+      .body(`Are you sure you want to delete this SWI?`)
+      .okBtn('Delete Stage')
+      .okBtnClass('btn btn-danger')
+      .cancelBtn('Cancel')
+      .cancelBtnClass('btn btn-secondary')
+      .open()
+      .then(dialogRef => dialogRef.result)
+      .then(result => {
+        //Delete logic goes here
+        this.swiFileService.deleteSWI(this.swi.id).then(((delSwi: SWIHeader) => {
+          this.toast.warning(this.swi.title + ' was deleted!', "Successfully Deleted");
+          this.navBack();
+        })).catch((err) => {
+          this.toast.error("Could not delete the SWI", "Delete failed");
+        });
+      })
+      .catch(err => console.log('Canceled'));
   }
 
 }
